refactor, not use finalize
